radiomanager-frontend: Introduce ChannelPage store hook
