+++ services/radiomanager-frontend/src/components/entries/ChannelPage/ChannelPage.tsx
@@ -1,22 +1,15 @@
 'use client'
 
-import { useCallback, useEffect, useState } from 'react'
 import { User, UserChannel, UserChannelTrack } from '@/api/api.types'
 import { Header } from '@/components/Header'
 import { Sidebar } from '@/components/Sidebar'
 import { StreamOverlay } from '@/components/StreamOverlay'
 import { LibraryLayout } from '@/components/layouts/LibraryLayout'
-import { ChannelTracksList, toChannelTrackEntry } from './ChannelTracksList'
+import { ChannelTracksList } from './ChannelTracksList'
 import { ChannelControls } from './ChannelControls'
-import { NowPlayingProvider, useNowPlaying } from '@/modules/NowPlaying'
-import {
-  deleteTracksById,
-  getChannelTracks,
-  MAX_TRACKS_PER_REQUEST,
-  removeTracksFromChannelById,
-} from '@/api/api.client'
-import { useMediaUploader, MediaUploaderComponent } from '@/modules/MediaUploader'
-import { UploadedTrackType } from '@/modules/MediaUploader/MediaUploaderTypes'
+import { NowPlayingProvider } from '@/modules/NowPlaying'
+import { MediaUploaderComponent } from '@/modules/MediaUploader'
+import { useChannelPageStore } from './hooks/useChannelPageStore'
 
 interface Props {
   channelId: number
@@ -31,82 +24,7 @@ export const ChannelPage: React.FC<Props> = ({
   userChannelTracks,
   userChannels,
 }) => {
-  const initialTrackEntries = userChannelTracks.map(toChannelTrackEntry)
-  const [trackEntries, setTrackEntries] = useState(initialTrackEntries)
-
-  const addTrackEntry = useCallback((track: UserChannelTrack) => {
-    setTrackEntries((entries) => [...entries, toChannelTrackEntry(track)])
-  }, [])
-
-  const initialCanInfinitelyScroll = initialTrackEntries.length === MAX_TRACKS_PER_REQUEST
-  const [canInfinitelyScroll, setCanInfinitelyScroll] = useState(initialCanInfinitelyScroll)
-
-  const handleInfiniteScroll = () => {
-    getChannelTracks(channelId, trackEntries.length).then((tracks) => {
-      const newEntries = tracks.map(toChannelTrackEntry)
-      setTrackEntries((entries) => [...entries, ...newEntries])
-
-      if (MAX_TRACKS_PER_REQUEST > newEntries.length) {
-        setCanInfinitelyScroll(newEntries.length === MAX_TRACKS_PER_REQUEST)
-      }
-    })
-  }
-
-  const { lastUploadedTrack } = useMediaUploader()
-  useEffect(() => {
-    if (!lastUploadedTrack) {
-      return
-    }
-
-    // The last uploaded track is not related to that channel
-    if (
-      lastUploadedTrack.type !== UploadedTrackType.CHANNEL ||
-      lastUploadedTrack.channelId !== channelId
-    ) {
-      return
-    }
-
-    // We haven't scrolled to the end of the tracks list
-    if (canInfinitelyScroll) {
-      return
-    }
-
-    addTrackEntry(lastUploadedTrack.track)
-  }, [lastUploadedTrack, addTrackEntry, canInfinitelyScroll, channelId])
-
-  const { refresh: refreshNowPlaying } = useNowPlaying()
-
-  const handleDeletingTracks = (trackIds: readonly number[]) => {
-    const idsSet = new Set(trackIds)
-    const updatedTrackEntries = trackEntries.filter(({ trackId }) => !idsSet.has(trackId))
-
-    setTrackEntries(updatedTrackEntries)
-
-    deleteTracksById(trackIds)
-      .then(() => {
-        refreshNowPlaying()
-      })
-      .catch((error) => {
-        // Restore tracks after unsuccessful delete
-        setTrackEntries(trackEntries)
-      })
-  }
-
-  const handleRemovingTracksFromChannel = (uniqueIds: readonly string[]) => {
-    const idsSet = new Set(uniqueIds)
-    const updatedTrackEntries = trackEntries.filter(({ uniqueId }) => !idsSet.has(uniqueId))
-
-    setTrackEntries(updatedTrackEntries)
-
-    removeTracksFromChannelById(uniqueIds, channelId)
-      .then(() => {
-        refreshNowPlaying()
-      })
-      .catch((error) => {
-        // Restore tracks after unsuccessful delete
-        setTrackEntries(trackEntries)
-      })
-  }
+  const channelPageStore = useChannelPageStore(channelId, userChannelTracks)
 
   return (
     <>
@@ -116,11 +34,11 @@ export const ChannelPage: React.FC<Props> = ({
         content={
           <ChannelTracksList
             channelId={channelId}
-            tracks={trackEntries}
-            canInfinitelyScroll={canInfinitelyScroll}
-            onInfiniteScroll={handleInfiniteScroll}
-            onDeleteTracks={handleDeletingTracks}
-            onRemoveTracksFromChannel={handleRemovingTracksFromChannel}
+            tracks={channelPageStore.trackEntries}
+            canInfinitelyScroll={channelPageStore.canInfinitelyScroll}
+            onInfiniteScroll={channelPageStore.handleInfiniteScroll}
+            onDeleteTracks={channelPageStore.handleDeletingTracks}
+            onRemoveTracksFromChannel={channelPageStore.handleRemovingTracksFromChannel}
           />
         }
         rightSidebar={
